third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/script-tag-replaced-with-questionmark-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?q=%3Cscript%3Ealert(%22xxx%fexxx%22)%3C/script%3E' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
